+++ forui/test/src/widgets/time_picker/time_picker_test.dart
@@ -125,4 +125,98 @@ void main() {
 
     expect(controller.value, const FTime());
   });
+
+  // The onChange callback may be called more than once for each change.
+  group('onChange', () {
+    testWidgets('when controller changes but onChange callback is the same', (tester) async {
+      int count = 0;
+      void onChange(FTime _) => count++;
+
+      final firstController = autoDispose(FTimePickerController(initial: const FTime(10, 30)));
+      await tester.pumpWidget(TestScaffold.app(child: FTimePicker(controller: firstController, onChange: onChange)));
+
+      firstController.value = const FTime(11, 45);
+      await tester.pump();
+
+      expect(count, 3);
+
+      final secondController = autoDispose(FTimePickerController(initial: const FTime(12)));
+      await tester.pumpWidget(TestScaffold.app(child: FTimePicker(controller: secondController, onChange: onChange)));
+
+      firstController.value = const FTime(13, 15);
+      secondController.value = const FTime(14, 30);
+      await tester.pump();
+
+      expect(count, 10);
+    });
+
+    testWidgets('when onChange callback changes but controller is the same', (tester) async {
+      int first = 0;
+      int second = 0;
+
+      final controller = autoDispose(FTimePickerController(initial: const FTime(10, 30)));
+      await tester.pumpWidget(TestScaffold.app(child: FTimePicker(controller: controller, onChange: (_) => first++)));
+
+      controller.value = const FTime(11, 45);
+      await tester.pump();
+
+      expect(first, 3);
+
+      await tester.pumpWidget(
+        TestScaffold.app(child: FTimePicker(controller: controller, onChange: (_) => second++)),
+      );
+
+      controller.value = const FTime(12);
+      await tester.pump();
+
+      expect(first, 3);
+      expect(second, 4);
+    });
+
+    testWidgets('when both controller and onChange callback change', (tester) async {
+      int first = 0;
+      int second = 0;
+
+      final firstController = autoDispose(FTimePickerController(initial: const FTime(10, 30)));
+      await tester.pumpWidget(
+        TestScaffold.app(child: FTimePicker(controller: firstController, onChange: (_) => first++)),
+      );
+
+      firstController.value = const FTime(11, 45);
+      await tester.pump();
+
+      expect(first, 3);
+
+      final secondController = autoDispose(FTimePickerController(initial: const FTime(12)));
+      await tester.pumpWidget(
+        TestScaffold.app(child: FTimePicker(controller: secondController, onChange: (_) => second++)),
+      );
+
+      firstController.value = const FTime(13, 15);
+      secondController.value = const FTime(14, 30);
+      await tester.pump();
+
+      expect(first, 3);
+      expect(second, 7);
+    });
+
+    testWidgets('disposed when controller is external', (tester) async {
+      int count = 0;
+
+      final controller = autoDispose(FTimePickerController(initial: const FTime(10, 30)));
+      await tester.pumpWidget(TestScaffold.app(child: FTimePicker(controller: controller, onChange: (_) => count++)));
+
+      controller.value = const FTime(11, 45);
+      await tester.pump();
+
+      expect(count, 3);
+
+      await tester.pumpWidget(TestScaffold.app(child: const SizedBox()));
+
+      controller.value = const FTime(12);
+      await tester.pump();
+
+      expect(count, 3);
+    });
+  });
 }